Update dependencies and refactor Makefile
- Update lezer dependencies to +1.0
- Use variables to organize prerequisites
- Use less phony targets
- Rename files
- Suggest using `npm install` instead of `yarn`
- Update Mocha and format `test-julia.js`
