+++ Makefile
@@ -3,25 +3,35 @@ ifdef DEBUG
 	LEZER_ARGS := --names $(LEZER_ARGS)
 endif
 
-.DELETE_ON_ERROR:
+# Written by humans
+SRC = src/julia.grammar src/tokens.js
 
-.PHONY: default
-default: build
+# Generated by lezer
+PARSER = src/_parser.js src/_parser.terms.js
 
-.PHONY: build
-build: dist/index.cjs dist/index.es.js
+# Bundled by rollup
+INDEX = dist/index.js dist/index.cjs
 
-.PHONY: test
-test:
+
+build: $(INDEX)
+$(INDEX): $(PARSER)
+	ROLLUP_IN="$(<)" \
+	ROLLUP_OUT="dist/index" \
+	npx rollup -c
+
+$(PARSER): $(SRC)
+	npx lezer-generator $(LEZER_ARGS) $(<) -o src/_parser
+
+
+.PHONY: test check clean
+
+test: $(INDEX) test/*
 	npx mocha test/test-julia.js
 
-.PRECIOUS: src/%.js src/%.terms.js
-src/%.js src/%.terms.js: src/%.grammar
-	@echo 'Compiling $(<)'
-	npx lezer-generator $(LEZER_ARGS) $< -o $(<:%.grammar=%)
+check: # Ensure tooling is installed
+	node --version
+	npm --version
 
-dist/%.cjs dist/%.es.js: src/%.js src/%.tokens.js src/%.terms.js
-	@echo 'Bunding $(<)'
-	ROLLUP_IN="$(<)" \
-	ROLLUP_OUT="$(<:src/%.js=dist/%)" \
-	npx rollup -c
+clean: # Remove generated files
+	rm $(PARSER)
+	rm $(INDEX)
