+++ bin/parse.js
@@ -3,7 +3,7 @@
 import { inspect } from "util";
 import * as fs from "fs";
 import { Tree } from "@lezer/common";
-import { parser } from "../src/index.js";
+import { parser } from "../src/_parser.js";
 
 function printTree(tree, input, from = 0, to = input.length) {
   let out = "";
